backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
requests==2.31.0
python-dotenv==1.0.0
